requirements.txt
aiofiles==0.4.0
aiohttp==3.5.4
aiohttp-index==0.1
aiohttp-jinja2==1.1.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
base58==1.0.3
chardet==3.0.4
idna==2.8
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.0.0
multidict==4.5.2
pluggy==0.7.1
py==1.8.0
pytest==3.6.4
pytest-asyncio==0.10.0
pytest-runner==4.4
python-box==3.3.0
python3-indy==1.8.0
serpy==0.3.1
six==1.12.0
toml==0.10.0
transitions==0.6.9
yarl==1.3.0
